+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Type a new commit message at the top of the list
 - Improve `fixup!` functions: create and apply
+- Reword the selected commit message with `r`
 
 ## [0.2.0][] - 2026-09-22